chore(db): env compare_server_default, %(here)s script_location, subprocess cwd, settings-driven test URLs
- alembic/env.py: compare_server_default=True in both configure() calls;
  alembic check reports no drift against the models.
- alembic.ini: script_location = %(here)s/alembic so alembic resolves the
  scripts dir regardless of cwd.
- tests: alembic subprocesses run with cwd=backend/; test_engine.py reads
  get_settings().test_database_url instead of hard-coded URLs.
- AiDebate.cost_usd annotated Mapped[Decimal | None] (Numeric column).
- DATABASE.md: updated_at ORM-only note also covers games.updated_at.

+++ backend/src/ffh/db/models/decisions.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import Any
 
 from sqlalchemy import (
@@ -72,7 +73,7 @@ class AiDebate(Base):
     consensus_score: Mapped[float] = mapped_column(REAL, nullable=False)
     disagreement_axis: Mapped[str | None] = mapped_column(Text)
     latency_ms: Mapped[int] = mapped_column(Integer, nullable=False)
-    cost_usd: Mapped[float | None] = mapped_column(Numeric(10, 6))
+    cost_usd: Mapped[Decimal | None] = mapped_column(Numeric(10, 6))
     tokens_in: Mapped[int | None] = mapped_column(Integer)
     tokens_out: Mapped[int | None] = mapped_column(Integer)
     cache_hit: Mapped[bool | None] = mapped_column(Boolean)
